split up packages, create seperate public models and fallout refactor
- moved bson handlign to its own package, and added more generic
  unmarshalling facility
- created public models for op types that users might interact with and
  implemented fmt.Stringer for optype
- implemented a Message -> models (as appropriate) converter.
- removed up wire.go file, pulling out header, optypes and messages.

+++ mongowire/wire_kill_cursors.go
@@ -8,7 +8,7 @@ func NewKillCursors(ids ...int64) Message {
 			RequestID: 19,
 			OpCode:    OP_KILL_CURSORS,
 		},
-		NumCursors: int64(len(ids)),
+		NumCursors: int32(len(ids)),
 		CursorIds:  ids,
 	}
 }
